packages/cli/python-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-wrap"
version = "0.1.0"
description = "Convert tool definitions from OpenAI, LangChain, and Claude formats to MCP-compatible YAML and generate MCP servers"
readme = "mcp_wrap/README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "langchain>=0.3.0",
    "pydantic>=2.0.0",
    "fastmcp>=0.1.0",
]

[project.scripts]
mcp-wrap = "mcp_wrap.mcp_wrap:main"

[project.urls]
Homepage = "https://github.com/yourusername/mcp-wrap"
Repository = "https://github.com/yourusername/mcp-wrap"
Issues = "https://github.com/yourusername/mcp-wrap/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_wrap*"]

[tool.setuptools.package-data]
mcp_wrap = ["demo/*", "README.md"] 